feat(prd-admin): 知识库多处入口 + 首页静态入口 SSOT 化
- 抽取首页静态入口（智能体/实用工具/基础设施）到 lib/homeLauncherItems，桌面首页 AgentLauncherPage 改为消费它，输出不变
- 移动端首页补回「基础设施」区（知识库置首），数据源与桌面首页统一，修复知识库在手机首页找不到的割裂
- 移动端「我的/设置」页新增「平台能力」快捷入口（知识库置首）
- 三处入口（首页/我的/抽屉）同源，名称顺序权限门一致，改一处处处同步
