+++ changelogs/2026-06-22_knowledge-base-multi-entry.md
@@ -1,0 +1,3 @@
+| refactor | prd-admin | 抽取首页静态入口（智能体/实用工具/基础设施）到 lib/homeLauncherItems 作为桌面+移动共用 SSOT |
+| feat | prd-admin | 移动端首页补回「基础设施」区（知识库置首），与桌面首页同源，消除知识库手机端找不到的割裂 |
+| feat | prd-admin | 移动端「我的/设置」页新增「平台能力」快捷入口（知识库置首），知识库多处可达 |
