+++ prd-admin/src/pages/MobileHomePage.tsx
@@ -10,6 +10,7 @@ import {
 } from 'lucide-react';
 import { accentFor, iconFor, DEFAULT_ACCENT } from '@/lib/agentAccent';
 import { useAuthStore } from '@/stores/authStore';
+import { buildStaticInfra, deriveLauncherPerms } from '@/lib/homeLauncherItems';
 import { BUILTIN_TOOLS } from '@/stores/toolboxStore';
 import type { ToolboxItem } from '@/services/real/aiToolbox';
 import { getAdminNotifications, getMobileFeed, getMobileStats } from '@/services';
@@ -56,6 +57,7 @@ const AGENT_EYEBROW: Record<string, { label: string; color: string }> = {
 export default function MobileHomePage() {
   const navigate = useNavigate();
   const cdnBase = useAuthStore((s) => s.cdnBaseUrl ?? '');
+  const permissions = useAuthStore((s) => s.permissions ?? []);
   const [notifications, setNotifications] = useState<AdminNotificationItem[]>([]);
   const [feed, setFeed] = useState<FeedItem[]>([]);
   const [stats, setStats] = useState<MobileStats | null>(null);
@@ -90,6 +92,8 @@ export default function MobileHomePage() {
   /* ── 数据分类 ── */
   const agents = useMemo(() => BUILTIN_TOOLS.filter((t) => t.kind === 'agent'), []);
   const tools = useMemo(() => BUILTIN_TOOLS.filter((t) => t.kind === 'tool'), []);
+  // 基础设施（知识库 / 资源 / 市场 / 工作流 / 更新中心 …）—— 与桌面首页同一数据源，知识库置首
+  const infra = useMemo(() => buildStaticInfra(deriveLauncherPerms(permissions)), [permissions]);
 
   /* ── 推荐 Carousel 项（有视频资产优先，无资产不上榜） ── */
   const featuredItems: FeaturedItem[] = useMemo(() => {
@@ -175,6 +179,24 @@ export default function MobileHomePage() {
                 subtitle: a.description,
                 tag: compatTagFor(a),
                 onClick: () => handleItemClick(a),
+              }))}
+            />
+          </AppStoreSection>
+        )}
+
+        {/* ── 基础设施（平台能力，知识库置首；与桌面首页同源，避免两端割裂） ── */}
+        {infra.length > 0 && (
+          <AppStoreSection title="基础设施" caption="知识库 · 资源 · 市场 · 工作流 · 更新中心">
+            <AppStoreRankedList
+              numbered={false}
+              items={infra.map((it) => ({
+                key: it.id,
+                Icon: iconFor(it.icon),
+                accent: accentFor(it.agentKey),
+                title: it.name,
+                subtitle: it.description,
+                pillLabel: '打开',
+                onClick: () => (it.routePath ? navigate(it.routePath) : undefined),
               }))}
             />
           </AppStoreSection>
